requirements.txt
torch
torchaudio
torchvision
torchdiffeq>=0.2.3
torchsde>=0.2.6
torchvision
einops>=0.6.0
open-clip-torch>=2.16.0
transformers>=4.29.1
safetensors>=0.3.0
pytorch-lightning>=2.0.0
aiohttp>=3.8.4
accelerate>=0.18.0
pyyaml>=6.0
scikit-image>=0.20.0
jsonmerge>=1.9.0
clean-fid>=0.1.35
clip @ git+https://github.com/openai/CLIP.git@main#egg=clip
resize-right>=0.0.2
opencv-python>=4.7.0.72
albumentations>=1.3.0
aiofiles>=23.1.0
frozendict>=2.3.6
python-dateutil>=2.8.2
importlib_resources
Pillow
scipy
tqdm
protobuf==3.20.3
psutil
ConfigArgParse
aio-pika